aws-backup-failed-monitoring/requirements.txt
boto3
PyYAML
xlsxwriter